Use hooks table as default
